валидация перенесена в сервисный слой из репозитория для item
